Update Chapter model and update course controller
Change the model of the course to set the number of chapters to 0 by default. Also change POST /api/courses and PUT /api/cporses to not require numberOfChapters in the body so that numberOfChapters changes separatly according to add and removed chapters. And add middlewared for chapter model to auto increment and decrement numberOfChapters in the course when adding and removing chapter